ocr/requirements.txt
Flask
Pillow
pytesseract
requests
python-dotenv